completed

+++ client/src/api/axios.ts
@@ -15,6 +15,11 @@ type ConflictResponseType = {
     status: ConflictType
   } 
 }
+type UserResponseType = { 
+  DATA : {
+    data: UserObjType
+  } 
+}
 export const BASEURL = 'http://localhost:5000/api/v1'
 
 export const contactApi = axios.create({
@@ -24,29 +29,33 @@ export const contactApi = axios.create({
   }
 })
 
-export const contactEndpoint = '/'
+export const getStarted = async (newUser: UserObjType): Promise<UserObjType> => {
+  const response = await contactApi.post('/user/get_started', newUser) as { data: UserResponseType };
+  return response.data.DATA.data
+}
 
+// contacts
 export const getContacts = async (): Promise<ContactObjType[]> => {
-  const response = await contactApi.get(contactEndpoint) as { data: ResponseType };
+  const response = await contactApi.get('/') as { data: ResponseType };
   return response.data.DATA.data
 }
 export const getConflict = async (email: string): Promise<ConflictType> => {
   const response = await contactApi.get(`/${email}`) as { data: ConflictResponseType };
   return response.data.DATA.status
 }
-export const addContact = async (newContact: Partial<ContactObjType>): Promise<ContactObjType> => {
-  const response = await contactApi.post('/create', newContact) as { data: SingleResponseType };
+export const addContact = async (newContact: Partial<ContactObjType>, userId: string): Promise<ContactObjType> => {
+  const response = await contactApi.post(`/create/${userId}`, newContact) as { data: SingleResponseType };
   return response.data.DATA.data
 }
 export const viewContact = async (contactId: string): Promise<ContactObjType> => {
   const response = await contactApi.patch(`/view_contact/${contactId}`) as { data: SingleResponseType };
   return response.data.DATA.data
 }
-export const updateContact = async (updatedContact: ContactObjType) => {
-  const response = await contactApi.put(`/update_contact`, updatedContact) as { data: SingleResponseType };
+export const updateContact = async (updatedContact: ContactObjType, userId: string) => {
+  const response = await contactApi.put(`/update_contact/${userId}`, updatedContact) as { data: SingleResponseType };
   return response.data.DATA.data
 }
-export const deleteContact = async (contactId: string) => {
-  return await contactApi.delete(`/delete/${contactId}`)
+export const deleteContact = async (contactId: string, userId: string) => {
+  return await contactApi.delete(`/delete/${contactId}/${userId}`)
   // return response.data
 }
